Initial Build
Created Unity3D project and assembled the primary game scene; created prefabs for the top and bottom walls, paddles, score zones, and the ball with relevant tags.  Wrote and implemented code for the ball, paddles, and game loop, resulting in a basic yet completely functional MVI.

Next Steps: Implement UI to keep track of scores and/or time limit and implement win conditions.