Added Special Balls and Pre-Game Pause
Added three special balls; the fast ball, which starts moving slow but gets faster over time, the flashing ball, which periodically turns invisible, and the small ball, which is smaller than normal and colored similar to the background, making it hard to see.

Added a pause when the game initially loads to give players more time to react to the first ball.

Added UI to the options screen describing each ball and implemented functionality for the Enable Special Balls toggle.