Modified UI and Added Ability to Reset Ball if "Softlocked"
Modified the appearance of the options menu's dropdowns to make the ability to scroll more evident.  Added a resume button to the pause menu.

The ball can now be reset during gameplay by pressing the R key; mainly a debug option, not intended for normal play, but can be used if the ball is bouncing at an angle that puts the match in a stalemate.

Rebuilt main build.